+++ proofs/adapter/filesystem.php
@@ -5,7 +5,7 @@ use Innmind\Filesystem\{
     Adapter\Filesystem,
     Directory\Directory,
     File\File,
-    File\Content\None,
+    File\Content,
     CaseSensitivity,
 };
 use Innmind\Url\Path;
@@ -48,9 +48,9 @@ return static function() {
         function($assert) {
             $property = new Adapter\AddRemoveAddModificationsStillAddTheFile(
                 Directory::named('0')
-                    ->add($file = File::named('L', None::of()))
+                    ->add($file = File::named('L', Content::none()))
                     ->remove($file->name()),
-                File::named('l', None::of()),
+                File::named('l', Content::none()),
             );
 
             $path = \sys_get_temp_dir().'/innmind/filesystem/';