+++ src/File/Content/OfStream.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 namespace Innmind\Filesystem\File\Content;
 
 use Innmind\Filesystem\{
-    File\Content,
     Stream\LazyStream,
     Exception\FailedToLoadFile,
 };
@@ -19,9 +18,10 @@ use Innmind\Immutable\{
 };
 
 /**
+ * @internal
  * @psalm-immutable
  */
-final class OfStream implements Content
+final class OfStream implements Implementation
 {
     /** @var callable(): Readable */
     private $load;
@@ -57,17 +57,17 @@ final class OfStream implements Content
         return $this->lines()->foreach($function);
     }
 
-    public function map(callable $map): Content
+    public function map(callable $map): Implementation
     {
         return Lines::of($this->lines()->map($map));
     }
 
-    public function flatMap(callable $map): Content
+    public function flatMap(callable $map): Implementation
     {
         return Lines::of($this->lines())->flatMap($map);
     }
 
-    public function filter(callable $filter): Content
+    public function filter(callable $filter): Implementation
     {
         return Lines::of($this->lines()->filter($filter));
     }