+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - `Innmind\Filesystem\Name` constructor is now private, use `::of()` named constructor instead
 - `Innmind\Filesystem\File\File` constructor is now private, use `::of()` named constructor instead
+- `Innmind\Filesystem\File\Content` is now a final class and its different implementations are declared internal, use the `Content` named constructors instead
 
 ### Removed